Rename radio-icon.tsx
